+++ Ma.Services.Appointments/Data/ApplicationDbContext.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace Ma.Services.Appointments
+{
+    public class ApplicationDbContext : DbContext
+    {
+        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
+            : base(options)
+        {
+        }
+        
+        public DbSet<AppointmentItem> Appointments { get; set; }
+    }
+}