+++ Ma.Services.Appointments/Data/MappingProfile.cs
@@ -1,0 +1,12 @@
+﻿using AutoMapper;
+
+namespace Ma.Services.Appointments
+{
+    public class MappingProfile : Profile
+    {
+        public MappingProfile()
+        {
+            CreateMap<AppointmentItem, DTO.AppointmentItem>();
+        }
+    }
+}