+++ Ma.Services.Appointments/Startup.cs
@@ -1,0 +1,101 @@
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Swashbuckle.AspNetCore.Swagger;
+using Swashbuckle.AspNetCore.SwaggerGen;
+using System;
+using System.IO;
+using System.Linq;
+using System.Reflection;
+
+namespace Ma.Services.Appointments
+{
+    public class Startup
+    {
+        public Startup(IConfiguration configuration)
+        {
+            Configuration = configuration;
+        }
+
+        public IConfiguration Configuration { get; }
+
+        // This method gets called by the runtime. Use this method to add services to the container.
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddDbContext<ApplicationDbContext>(options => options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection")));
+
+            services.AddApiVersioning();
+            services.AddAutoMapper(typeof(Startup));
+            
+            services.AddSwaggerGen(c =>
+            {
+                c.SwaggerDoc("v1.0", new Info
+                {
+                    Title = "Appointment Management API",
+                    Version = "v1.0",
+                    Description = "The goal of the appointment service is to provide a cloud-based functionallity to help you to organize you time effective.",
+                    Contact = new Contact
+                    {
+                        Name = "Alexandr Miller",
+                        Email = "info@miller-alex.de"
+                    },
+                    License = new License
+                    {
+                        Url = "https://raw.githubusercontent.com/miller-alex-github/personal-site/master/LICENSE"
+                    }
+                });
+
+                c.EnableAnnotations();
+                c.DocInclusionPredicate((docName, apiDesc) =>
+                {
+                    if (!apiDesc.TryGetMethodInfo(out MethodInfo methodInfo)) return false;
+
+                    var versions = methodInfo.DeclaringType
+                        .GetCustomAttributes(true)
+                        .OfType<ApiVersionAttribute>()
+                        .SelectMany(attr => attr.Versions);
+
+                    return versions.Any(v => $"v{v.ToString()}" == docName);
+                });
+
+                // XML Documentation
+                var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+                var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+                c.IncludeXmlComments(xmlPath);
+            });
+
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+        }
+
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+            else
+            {
+                app.UseHsts();
+            }
+            
+            app.UseHttpsRedirection();
+            app.UseMvc();
+            app.UseSwagger(c =>
+            {
+                c.RouteTemplate = "docs/{documentName}/docs.json";
+            });
+            app.UseSwaggerUI(c =>
+            {
+                c.SwaggerEndpoint("/docs/v1.0/docs.json", "Appointment Management API - v1.0");
+                c.RoutePrefix = "docs";
+            });
+
+            app.AutoMigrateDatabase();
+        }
+    }
+}