+++ Ma.Web/Migrations/20191110171715_InitialCreate.cs
@@ -2,9 +2,9 @@
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Migrations;
 
-namespace Ma.Web.Data.Migrations
-{   
-    public partial class CreateIdentitySchema : Migration
+namespace Ma.Web.Migrations
+{
+    public partial class InitialCreate : Migration
     {
         protected override void Up(MigrationBuilder migrationBuilder)
         {