+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [1.2] In process
 - Redesign main layout of footer.
-- Create new appointment table in database.
+- Add micro service for appointments (https://appointments.miller-alex.de/docs/index.html).
+- Added unit test scaffold
+- 
 
 ## [1.1] 2019-11-05
 - Redesign the main layout of the page (logo, menu).